+++ auth-service/src/main/java/com/klp/authservice/auth/presentation/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.klp.authservice.auth.entrypoint.dto.response;
+package com.klp.authservice.auth.presentation.dto.response;
 
 public record LoginResponse(
     Long userId,
