+++ auth-service/src/main/java/com/klp/authservice/auth/presentation/dto/response/ReissueResponse.java
@@ -1,4 +1,4 @@
-package com.klp.authservice.auth.entrypoint.dto.response;
+package com.klp.authservice.auth.presentation.dto.response;
 
 public record ReissueResponse(
     Long userId,
